* Breakup the Bson.cs file.
