+++ Source/Bogus/Bson/BValueType.cs
@@ -1,0 +1,16 @@
+namespace Bogus.Bson
+{
+   public enum BValueType
+   {
+      Double,
+      String,
+      Array,
+      Binary,
+      Boolean,
+      UTCDateTime,
+      None,
+      Int32,
+      Int64,
+      Object
+   }
+}
